+++ lib/glimmer/libui/square.rb
@@ -23,16 +23,11 @@ require 'glimmer/libui/shape'
 
 module Glimmer
   module LibUI
-    class Arc < Shape
-      parameters :x_center, :y_center, :radius, :start_angle, :sweep, :is_negative
-              
+    class Square < Shape
+      parameters :x, :y, :length
+    
       def draw(area_draw_params)
-        @args[5] ||= ControlProxy.boolean_to_integer(@args[5], allow_nil: false)
-        if parent.is_a?(Figure) && parent.x.nil? && parent.y.nil?
-          ::LibUI.draw_path_new_figure_with_arc(path_proxy.libui, *@args)
-        else
-          ::LibUI.draw_path_arc_to(path_proxy.libui, *@args)
-        end
+        ::LibUI.draw_path_add_rectangle(path_proxy.libui, *@args, length)
         super
       end
     end